Extend occurs check to cover regressions.
Also add unit test suite to inference (ported from chalk) so I can write
a test. It occurs to me that this version of the occurs check is a
conservative one -- it prevents a variable from being bound to a
projection from itself, which may be stricter than necessary, but seems
reasonable. If we didn't do this, we would need some add'l conditions to
prevent infinitely sized types, I believe.

Fixes #5.