Separate Program from ProgramEnvironment.
Closes #19. I took a slightly different approach from your suggestion.

ir::Program and ProgramEnvironment are distinct types with no ownership
relationship, but instead of lowering returning a tuple, ir::Program
has a method to construct a ProgramEnvironment.

In addition to the trait_data, the associated_type_data also needs to
be cloned over, and the `split_projection` method is duplicated for
both types, because that method is invoked in `elaborated_clauses`.

It seems like once we have a new way of solving #12, ProgramEnvironment
would just be a set of ProgramClauses, and that would be the whole
context necessary for the solver to run.