First version of overlap check.
This check works by pairing every 2 impls of the same trait and then
generating a goal to prove that a type lies in the intersection of
those two impls. If the goal succeeds, we have found the overlap.

The goal that is lowered looks like this:

```rust
impl Foo for i32 { }
impl Foo for u32 { }

exists<> { i32 = u32 }
```

```rust
impl<T> Foo for T { }
impl Foo for i32 { }
exists<T> { T = i32 }
```

```rust
impl<T> Foo for Vec<T>
impl<T> Foo for Vec<i32>
exists<T> { Vec<T> = Vec<i32>
```

And so on.

Currently, this goal does not take into consideration where clauses
on the impl. For that reason, the local negative reasoning, as in the
example below, is not implemented:

```rust
trait Bar { }
trait Foo { }
struct MyType;
impl<T: Bar> Foo for T { }
impl T for MyType { }
```

A test exists to check this, which is failing.