Add a generalized folder trait (incomplete)
A folder trait GeneralFolder is added that has functions to fold types
and lifetimes. A bridge impl is added that maps the functions of the
GeneralFolder trait to the functions of the existing Folder trait. One
mapping is still missing: the author does not know how to map the
Lifetime::ForAll(..) variant to the fold_free_lifetime_var(..) function.