remove `cannot_prove` from unify, change approach to universals
Currently unification does not reject `i32 = T` (where `T` is a
universally quantified variable). This leads to a fair amount of
complexity, since we must now distinguish another kind of solution
("cannot prove") which is "similar but different" to ambiguity.

This commit removes all of that logic in favor of introducing a refined
concept of the "must be ground" conditions from the negation-as-failure
paper. In particular, the original paper noted that free inference
variables wind up acting as universal constraints when negated -- which
is exactly what we want here. So we introduce a new function `negated`
that transforms a goal G by (a) forbidding free existential variables
and (b) converting free universal variables into existential ones. If
the resulting query G' is satisfiable, then `not { G }` is false.