replace fallback clauses with `ProjectionEq`/`Normalize`
Instead of having "fallback clauses" to handle projection, we now
have `ProjectionEq(<T as Trait>::Bar = U)`, which can be proven
either by normalizing *or* by skolemizing.

This required adjusting how we handle elaboration. That is, we now
"expand" a `T: Trait<Bar = U>` (which is the user syntax for projection
equality) into `T: Trait` as well, which in turn expands into `WF(T:
Trait)`. We do this instead of having a clause like:

> `(T: Trait) :- exists<U> { (T: Trait<Bar = U>) }`

This is because that clause caused issues for the recursive solver; in
particular, `exists<U> { (T: Trait<Bar = U>) }` always resolves to an
ambiguous result, since there is no unique `U`.

This PR adjusts the tests appropriately. Mostly we add normalize and
projection. In a few cases, the SLG solver doesn't work yet; those are
tagged with FIXME (and fixed in a later commit).