substitute answers into pending goals without using unification
This is not only more efficient, but it also sidesteps the problems that
led to the SLG solver recursing infinitely when unifying goals involving
projections.

This also uncovers a potential flaw with the truncation operation, in
that it does not concern itself with internal binders. For example, we
might truncate a type like `for<'a> fn(Vec<&'a u32>)` into `for<'a>
fn(?X)`. This is problematic because there is no value of `?X` that
could reference `'a`. For now, the SLG solver will panic in this
scenario, but probably truncation needs to truncate the binder as well.