pass `&impl ContextOps` as parameter
We used to to store a `impl ContextOps` in `Forest`, but part of the
job of the `ContextOps` is going to be to consult the database and do
queries. Query implementations only have borrowed access to the salsa
database, so that means we need to take the `ContextOps` as a
parameter.