remove unselected projection types
The concept of an "unselected projection" was meant to model the Rust
syntax like `T::Item`, where the trait is unspecified. It's meant to
be a more principled version of the ad-hoc logic that rustc presently
uses. This is still a good idea, but for now it's not worth dealing
with, and it makes the code more complex.