WIP: fix substitute to return the data within, substituted
that however leads us to have to figure out how to handle unifying
`forall<'a> T` with some other type `U`. But in practice, in Rust,
forall types will *only* be created for function types, and hence such
unification never makes sense. So that's the semantics I implemented,
but it is causing some of the tests to "look funny", because now the
syntax for `fn` in chalk is `for<>`, effectively. =)