permit inference vars to be unified with non-normalized projections
I have a dim memory that this violates some invariant I wanted to
maintain, but I can't right now be sure that that is. I suppose it
means that answers can contain unnormalized projections, in theory,
which we might want to prevent.

The only time this is needed is when equating answers after
truncation. It's a rather special case and we might be able to do a
"smaller" fix.