track outermost binder instead of number of binders
This is more strongly typed. In the process, rename `shifted_in_by/shifted_out_by` to

* `shifted_in_from(D)` -- i.e., we are shifting a value from the binder D into this one
* `shifted_out_to(D)` -- i.e., we are shifting a value from this binder into the binder D

I at least find those names give me a better intuition for what's going on in
most cases. i.e., we are taking a term that was valid from some outer binder D
and shifting it *in* to the inner binder.