Make it possible to cache the result of env elaboration
`program_clauses_for_env` can take a huge amount of time -- rust-analyzer is
unusably slow on the Chalk codebase itself, spending most of that time in env
elaboration because one `I: Interner` results in ~300 clauses, which gets
repeated tens of times per query.

This adds a method to `RustIrDatabase` that should basically just be a
pass-through for `program_clauses_for_env`, but which makes it possible to cache
the results in Salsa.

The effect of this is a ~20% speedup of analysis of rust-analyzer on itself, and
turning analysis of the Chalk codebase from seemingly forever to ~1 minute.