Fix associated ty parameter names
AssociatedTyDatum's parameters are created inside of a new enviroment,
and include the surrounding trait implementation's parameters.

This splits the trait's parameters, and then records their names in the
trait's environment, and their name in the assoc ty's environment in
a mapping table in WriterState. This table is then passed down, and
whenever parameter names are needed, their assoc ty names are generated
and then their names in the trait environment are retreived from this
table.

Co-authored-by: David Ross <daboross@daboross.net>