Handle bound vars in builtin trait impls
The builtin impls have the same problem that the dyn trait impl did, which is
that they reuse the trait ref as given. The problem with that is that the
recursive solver calls `program_clauses` with trait refs that contain bound vars
from the original goal. So the builtin impls returned clauses with free bound
vars, which is a problem. The solution is the same as for the dyn trait impls:
Use the `Generalize` folder to collect those free bound vars and basically put a
`forall` around them. To give an example:

Looking for

```
Implemented((^0.0, ^0.1): Sized)
```

previously returned something like

```
if (Implemented(^0.1: Sized)) { Implemented((^0.0, ^0.1): Sized) }
```

as a clause; now it returns

```
forall<type, type> if (Implemented(^0.1: Sized)) { Implemented((^0.0, ^0.1): Sized) }
```

(The SLG solver instantiates goals before asking for program clauses, so it has
inference variables instead of bound vars in those places.)