store the "minimums" on the stack
We used to pass the `Minimums` structure as a parameter.  We now store
it in the `Stack` data structure, which is equivalent.  The reason for
this change is that it allows the solve/fulfill code to be entirely
ignorant of the Minimums tracking and, more generally, the logic around
re-execution of cyclic goals.