Fix program writer test item ordering
We test the program writer by parsing a program, writing it, and
reparsing the output. We compare the original Program against the
writer's output using Program's Eq impl. This Eq impl relies on item
ids, which are dependent on the order the items appear in the program
source.

Previously, we would output one type of item at a time. That meant the
original ordering would not be preserved, unless the program was written
with the items grouped, in the correct ordering.

This is a weird and frustrating quirk that makes the tests less
readable. We have updated `write_program` to preserve the ordering. The
items are sorted by their RawId, which will result in the original
ordering.

Co-authored-by: David Ross <daboross@daboross.net>