Refactor and document syntax writer tests
Refactored the syntax writer tests so there is only one reparse_test
call per test function. This involved removing reduntant tests, and
creating new test functions.

We also added comments for all of the test to describe what particular
functionality the test was targeted at. This is useful when a test's
purpose is obscured by the substantial boilerplate that test requires.

Co-authored-by: David Ross <daboross@daboross.net>