Add default `RustIrDatabase::name_*` methods
The .chalk syntax writer needs names for items such as adts,
opaque types, and traits. `RustIrDatabase::name_*` methods were added to
provide these names, but they increase the burden on implementors.

To remove this burden, we added default impelemtnations to
`RustIrDatabase` that sanitize the output of the `Interner::debug_*`
methods so they are only alphanumeric ascii characters.

We also removed the custom implementations for the .chalk syntax writer
tests to ensure that the default implementations are tested. We had to
keep the implementation for associated types because the debug methods
outputted different names. This broke the reparse test because the
`Program`s were no longer considered equal by the derived `Eq`
implementation.

Co-authored-by: David Ross <daboross@daboross.net>