Fix tls context in LoggingRustIrDatabase tests
The previous commit switched from custom `RustIrDatabase::name_*`
implementations that directly used the data in `Program` to using the
debug methods in the `Interner`.

This exposed a bug in the test harness for logging db. The debug methods
rely on thread local storage to get access to the data inside of
`Program`. The bug caused the program to run outside of the thread local
storage context that provided program, causing the names to change to a
fallback value. This broke the re-solve tests because the names needed
to match up with the goal.

This commit fixes the tests and underlying bug by correctly running
`LoggingRustIrDatabase::to_string` in the tls context.

Co-authored-by: David Ross <daboross@daboross.net>