Fix duplicate names across multiple writes
This fixes two bugs in unique name generation. The bug which prompted
this change is related to stub items: LoggingRustIrDatabase calls
`write_stub_items` and `write_items` separately, and unique name choices
were not persisted between these calls. The second bug, discovered by
our test framework, was that associated type names were not made unique
within a trait.

We refactored writer state into a public `WriterState`, and a private
`InternalWriterState`. The public WriterState holds a copy of the DB and
the name de-duplication information. The private `InternalWriterState`
holds the current debrujin index information and formatting data.

The public `WriterState` is created outside of the display code, and is
saved between writes. In the case of the LoggingDB, it is stored inside
the logging db. This ensures that all writes from the LoggingDB have
consistent unique names.

Storing `WriterState` persistently is not strictly necessary; to fix the
unique name choice bug, we only need to persist it between the calls to
`write_stub_items` and `write_items` in `Display::fmt`. But storing it
longer was easy to do, and not less correct.

Co-authored-by: David Ross <daboross@daboross.net>