Cleanup remnants of granular generalizer relate
In an eariler commit, we switched from relating each new inference var
with the corrosponding part of the type individually, to relating the
entire generalized type with the type we are relating with. This ensures
that when we traverse through bounders, the correct lifetime constraints
are created.

When we switched, we did not remove the error handling logic. This
commit finishes the switch be removing the commit.

Co-authored-by: David Ross <daboross@daboross.net>