Fix ignored var ty ordering in relate_ty_ty
relate_ty_ty was ignoring the order in which the var and ty were passed
in when matching a var and a ty. This mean't if the var was passed in
second, the variance would be incorrectly inverted because the
parameters were swapped around.

We fixed this by inverting the variance if the var is on the right side,
and added a test case for the issue.

We also had to switch the order we pass the value + generalized value
into relate_ty_ty, since the previous incorrect ordering was hidden by
the bug.

Co-authored-by: David Ross <daboross@daboross.net>