Fix incosistent usage of push_lifetime_outlives
The code that relates Adt's was using `push_lifetime_outlives_goals` in
a way that assumed it would push a outlives b when the variance is
covariant. The code that relates two references assumed that it would
push a constraint b outlives a when the variance was covariant. This
latter behavior was what the function actually did, despite its name
implying the opposite.

This commit changes the behavior of `push_lifetime_outlives_goals` to
push a outlives b when the variance is covariant, and updates the code
that relates references.

We also updated an incorrect test case where a reference was used in a
struct. This test case would have failed (because
`push_lifetime_outlives_goals` was inverted), but the constraints in the
expected output were also incorrectly inverted.

The output of two additional test cases also were updated because the
constraints were emitted in a different order after making the update.

Co-authored-by: David Ross <daboross@daboross.net>