Fix handling of variables in goal for new alias clauses
The new clauses for alias types (`<X as Y>::Z: Trait :- T: Trait, <X as Y>::Z ==
T` etc.) were generated using the provided goal, without regard for bound
variables. This led to e.g. the new test failing in the recursive solver.
Alternatively, it might be better to generate the clauses without using the
specific types in the goal.