Add binders validator
This is a visitor which checks that all bound vars refer to an existing
binder of the right type. This has been quite useful for me: we somewhat
regularly have errors in the binders which then cause a crash some time
down the line when something tries to substitute them. At that point
it's usually harder to find the actual place the error comes from; also
you don't always hit the error. For the latest of these bugs, I added
this validator, and immediately had failing tests that pointed me to the
right location.

The disadvantage is that I needed to add some rather specific methods to
`Visitor`. I didn't find a better way to model this (and this is also
the reason I couldn't just implement it in rust-analyzer).