Fix automatic changelog generation to include current version's changes
The changelog generation as it was written would not include the latest
"unreleased" changes since the git tag for the new release hadn't been
created yet. This change fixes that issue by using the
`--latest-version` flag of `auto-changelog`.

The `--latest-version` flag replaces the "Unreleased" section with the
given version, allowing us to generate a changelog before creating the
git tag it references. Ex. `--latest-version 0.70.0` will create a
`v0.70.0` entry in the changelog even though the latest tag is
`v0.69.0`.