Omit empty lists from `ConstrainedSubsts`
This is harder than it should be due to the way we write `Display`
impls. `ConstrainedSubsts` is wrapped in `Canonical`, a container over a
generic `T`, so the `Display` impl for `Canonical` cannot know that the
contained type has a `display` method that bundles the interner. As a
result, the interner is gone by the time we get to the `Display` impl
for `ConstrainedSubsts`.

I think a better solution is to implement a custom `DebugWith<Ctxt>`
trait (see Jonathan Turner's `lark` for prior art).